expgfx_addremove 88.11->88.43% ((int)slotType call-arg cast forces extsh-into-short-callee-param at expgfxGetSlot, #115)

+++ src/main/dll/dll_000A_expgfx.c
@@ -2801,7 +2801,7 @@ int expgfx_addremove(ExpgfxSpawnConfig* config, int preferredPoolIndex, short sl
     {
         return EXPGFX_INVALID_POOL_INDEX;
     }
-    if (expgfxGetSlot(&poolIndex, &slotIndex, slotType,
+    if (expgfxGetSlot(&poolIndex, &slotIndex, (int)slotType,
                       preferredPoolIndex, (u32)(int)config->attachedSource)
         == EXPGFX_INVALID_POOL_INDEX)
     {
